+++ jipp-core/src/test/kotlin/com/hp/jipp/dsl/DslTest.kt
@@ -1,15 +1,20 @@
 package com.hp.jipp.dsl
 
+import com.hp.jipp.encoding.AttributeGroup.Companion.groupOf
+import com.hp.jipp.encoding.AttributeGroup.Companion.mutableGroupOf
 import com.hp.jipp.encoding.Cycler.cycle
 import com.hp.jipp.encoding.IntOrIntRange
 import com.hp.jipp.encoding.IppPacket
 import com.hp.jipp.encoding.IppPacket.Companion.DEFAULT_REQUEST_ID
+import com.hp.jipp.encoding.KeywordOrName
 import com.hp.jipp.encoding.MediaSizes
 import com.hp.jipp.encoding.Name
 import com.hp.jipp.encoding.Tag
 import com.hp.jipp.model.BindingType
 import com.hp.jipp.model.Media
 import com.hp.jipp.model.MediaCol
+import com.hp.jipp.model.MediaSource
+import com.hp.jipp.model.MediaType
 import com.hp.jipp.model.Operation
 import com.hp.jipp.model.Status
 import com.hp.jipp.model.Types
@@ -206,5 +211,22 @@ class DslTest {
     fun `construct with operation`() {
         val packet = IppPacket.Builder(Operation.fetchJob).build()
         assertEquals(Operation.fetchJob, packet.operation)
+    }
+
+    @Test
+    fun `coerce types from mutableGroups`() {
+        val mediaType1 = MediaCol().apply {
+            mediaSize = MediaSizes.parse(Media.naLetter8p5x11in)
+            mediaLeftMargin = 750
+            mediaRightMargin = 750
+            mediaBottomMargin = 750
+            mediaTopMargin = 750
+            mediaSource = KeywordOrName(MediaSource.main)
+            mediaType = KeywordOrName(MediaType.stationery)
+        }
+        val packet = IppPacket(Status.successfulOk, 1234, groupOf(Tag.operationAttributes),
+            groupOf(Tag.printerAttributes, Types.mediaColReady.of(mediaType1)))
+        val printerAttributes = mutableGroupOf(Tag.printerAttributes, cycle(packet)[Tag.printerAttributes]!!)
+        assertEquals(listOf(mediaType1), printerAttributes.getValues(Types.mediaColReady))
     }
 }