+++ jipp-pdl/src/test/kotlin/pclm/PclmTest.kt
@@ -8,6 +8,7 @@ import com.hp.jipp.pdl.OutputSettings
 import com.hp.jipp.pdl.pclm.PclmSettings
 import com.hp.jipp.pdl.pclm.PclmWriter
 import java.io.ByteArrayOutputStream
+import java.io.File
 import org.junit.Test
 import util.ByteWindow
 import util.RandomDocument
@@ -15,7 +16,7 @@ import util.RandomDocument
 class PclmTest {
 
     @Test
-    fun validateGeneratedPclm() {
+    fun `validate PCLM`() {
         cyclePclm(PclmSettings(stripHeight = 32))
     }
 
@@ -26,11 +27,12 @@ class PclmTest {
 
     private fun cyclePclm(caps: PclmSettings) {
         // Use a tall enough page so that we're assured there will be at least one blank area
-        val randomDocument = RandomDocument(12345L, 2, 72.0, 150.0, 300)
+        val randomDocument = RandomDocument(12345L, 2, 283.46457, 419.52756, 300)
         val bytesOut = ByteArrayOutputStream()
         PclmWriter(bytesOut, caps).use {
             it.write(randomDocument)
         }
+        bytesOut.toByteArray().inputStream().copyTo(File(".", "testout.pclm.pdf").outputStream())
 
         // Ensure that the resulting PCLM follows all rules we have identified
         validatePclm(bytesOut.toByteArray())